edits

+++ testsuperpixel.cu
@@ -1,12 +1,12 @@
 
-__global__ void testsuperPixel(float *inputs, int *output)
+__global__ void testsuperPixel(int *inputs, int *output)
 {
 	int blockId = blockIdx.x + blockIdx.y * gridDim.x; 
 	int globalId = blockId * (blockDim.x * blockDim.y) + (threadIdx.y * blockDim.x) + threadIdx.x;
 	int localId = (threadIdx.y * blockDim.x) + threadIdx.x;              
 
 	// Initialize local sum array to be filled in with values from our input array
-	__shared__ float inputsToSum[16];
+	__shared__ int inputsToSum[16];
 
 	// Assign values from input value array to our local sum array
     inputsToSum[localId] = inputs[globalId];
